Refactor HeroSection into modular components
- Split the HeroSection into multiple components: `HeroBackground`, `HeroHeadline`, `HeroMapCard`, `HeroParallaxCard`, and `HeroStats` for improved organization and maintainability.
- Introduced a new `README.md` for documentation on the modular structure and usage.
- Enhanced the parallax effect and animations for a more dynamic user experience.
- Updated the main `HeroSection` to utilize the new components, ensuring a cleaner and more scalable codebase.
